20240519 컨트롤러 변경 반영

+++ src/main/java/com/example/demo/comment/service/mBoardCommentServiceImpl.java
@@ -3,7 +3,9 @@ package com.example.demo.comment.service;
 import java.util.ArrayList;
 import java.util.List;
 
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import com.example.demo.comment.dto.CommentDTO;
@@ -16,6 +18,9 @@ public class mBoardCommentServiceImpl implements mBoardCommentService {
 	
 	@Autowired
 	private mBoardCommentRepository repository;
+
+	@Autowired
+	AuthenticationFacade authenticationFacade;
 	
 	@Override
 	public int register(CommentDTO dto) {
@@ -41,7 +46,16 @@ public class mBoardCommentServiceImpl implements mBoardCommentService {
 
 	@Override
 	public void remove(int no) {
-		repository.deleteById(no);
+		mBoardComment comment = repository.findById(no)
+				.orElseThrow(()-> new EntityNotFoundException("해당 댓글을 찾지 못했습니다. "));
+
+		Authentication authentication = authenticationFacade.getAuthentication();   // 현재 로그인한 사용자 확인
+
+		if(!comment.getWriter().equals(authentication.getName())){
+			System.out.println("해당 댓글은 삭제 불가능합니다. (접근 불가능)");
+		}else{
+			repository.deleteById(no);
+		}
+	}
 	}	
 
-}
